# cmd/docs/swagger.yaml
basePath: /api
definitions:
  models.Star:
    properties:
      dec:
        type: number
      description:
        type: string
      imageURL:
        type: string
      isActive:
        type: boolean
      observations:
        description: связь многие-ко-многим через telescope_observation_stars
        items:
          $ref: '#/definitions/models.TelescopeObservation'
        type: array
      ra:
        type: number
      shortDescription:
        type: string
      starID:
        type: integer
      starName:
        type: string
    type: object
  models.TelescopeObservation:
    properties:
      completionDate:
        type: string
      createdAt:
        type: string
      creator:
        $ref: '#/definitions/models.User'
      creatorID:
        type: integer
      formationDate:
        type: string
      moderator:
        $ref: '#/definitions/models.User'
      moderatorID:
        type: integer
      observationDate:
        type: string
      observerLatitude:
        type: number
      observerLongitude:
        type: number
      stars:
        items:
          $ref: '#/definitions/models.Star'
        type: array
      status:
        type: string
      telescopeObservationID:
        type: integer
      telescopeObservationStars:
        items:
          $ref: '#/definitions/models.TelescopeObservationStar'
        type: array
    type: object
  models.TelescopeObservationStar:
    properties:
      orderNumber:
        type: integer
      quantity:
        type: integer
      resultValue:
        type: number
      star:
        $ref: '#/definitions/models.Star'
      starID:
        type: integer
      telescopeObservation:
        $ref: '#/definitions/models.TelescopeObservation'
      telescopeObservationID:
        type: integer
    type: object
  models.User:
    properties:
      isModerator:
        type: boolean
      passwordHash:
        type: string
      userID:
        type: integer
      username:
        type: string
    type: object
host: 192.168.1.51:9005
info:
  contact: {}
  description: "Система позволяет пользователям создавать заявки на получение данных
    для расчёта наведения телескопа,\nа модераторам — управлять этими заявками и добавлять
    новые звёзды.\n## \U0001F510 Аутентификация\n- Используются **сессии и cookie**,
    хранящиеся в **Redis**.\n- Без авторизации доступны только методы **чтения (GET)**.\n-
    Для авторизованных запросов браузер автоматически отправляет cookie.\n## \U0001F465
    Роли и права доступа\n- **Гость:** только GET-запросы.\n- **Пользователь:** управление
    своими заявками + чтение данных.\n- **Модератор:** полный доступ ко всем ресурсам."
  title: Calculator Observations Stars API
  version: "1.0"
paths:
  /stars:
    get:
      consumes:
      - application/json
      description: Возвращает список звёзд, доступных для наблюдения. Можно указать
        фильтр по названию.
      parameters:
      - description: Фильтр по названию звезды
        in: query
        name: star_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Star'
            type: array
        "500":
          description: Ошибка получения звёзд
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить список звёзд
      tags:
      - Stars
    post:
      consumes:
      - application/json
      description: Создаёт новую запись звезды в базе данных
      parameters:
      - description: Данные звезды
        in: body
        name: star
        required: true
        schema:
          $ref: '#/definitions/models.Star'
      produces:
      - application/json
      responses:
        "201":
          description: Звезда успешно добавлена
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Некорректный JSON или отсутствует название
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка при сохранении
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Добавить новую звезду
      tags:
      - Stars
  /stars/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет звезду по ID
      parameters:
      - description: ID звезды
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Звезда успешно удалена
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Некорректный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Звезда не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка при удалении
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить звезду
      tags:
      - Stars
    get:
      consumes:
      - application/json
      description: Возвращает информацию о конкретной звезде по её ID
      parameters:
      - description: ID звезды
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Star'
        "404":
          description: Звезда не найдена
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить звезду по ID
      tags:
      - Stars
    put:
      consumes:
      - application/json
      description: Обновляет информацию о звезде по ID
      parameters:
      - description: ID звезды
        in: path
        name: id
        required: true
        type: integer
      - description: Обновлённые данные звезды
        in: body
        name: star
        required: true
        schema:
          $ref: '#/definitions/models.Star'
      produces:
      - application/json
      responses:
        "200":
          description: Звезда успешно обновлена
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Некорректный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Звезда не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка при обновлении
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновить данные звезды
      tags:
      - Stars
  /stars/{id}/add:
    post:
      consumes:
      - application/json
      description: Добавляет звезду в текущую заявку пользователя (черновик). Если
        уже есть — увеличивает количество.
      parameters:
      - description: ID звезды
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Звезда успешно добавлена или количество увеличено
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Некорректный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка при работе с заявкой
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Добавить звезду в черновик заявки
      tags:
      - Stars
  /stars/{id}/image:
    post:
      consumes:
      - multipart/form-data
      description: Загружает изображение звезды в MinIO и сохраняет путь в БД
      parameters:
      - description: ID звезды
        in: path
        name: id
        required: true
        type: integer
      - description: Файл изображения
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Изображение успешно загружено
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Некорректный ID или файл
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка при загрузке или сохранении
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Загрузить изображение звезды
      tags:
      - Stars
...
securityDefinitions:
  SessionAuth:
    in: cookie
    name: session_token
    type: apiKey
swagger: "2.0"